Finished iric for delete

+++ ClassicCarsSpringBoot/src/main/java/com/vladolium/odinmodel/serviceImpl/CustomersServiceImpl.java
@@ -34,19 +34,26 @@ public class CustomersServiceImpl implements CustomersService {
 	}
 	
 	
-	
-	
-	
-	
-	
-	
-	
-	
-	
 	@Override
 	public Customers readOneById(Long id) {
 		return customersRepository.getOne(id);
 	}
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
 	
 	
 	
@@ -61,13 +68,6 @@ public class CustomersServiceImpl implements CustomersService {
 	
 	
 	
-	
-	
-	
-	
-	
-	
-	
 	@Override
 	public Iterable<Customers> readAll() {
 		return customersRepository.findAll();
@@ -77,10 +77,6 @@ public class CustomersServiceImpl implements CustomersService {
 	public Page<Customers> readAllPagination(Pageable page) {
 		return customersRepository.findAll(page);
 	}
-	
-	
-	
-	
 	
 	
 	
@@ -103,51 +99,55 @@ public class CustomersServiceImpl implements CustomersService {
 	
 	
 	
+	
+	
+	
+	
 	@Override
 	public Iterable<Customers> search(
 		
 		
-		
-		
-		
-		
 		Long employeesId,
 		
+		
+		
+		
+		
 		List<Reviews> reviewsList,
-		String postalCode,
-		String addressLine1,
 		String lastName,
-		Double creditLimit,
-		String firstName,
-		String country,
 		String customerName,
 		String phone,
 		String addressLine2,
+		Double creditLimit,
+		String addressLine1,
+		String firstName,
+		String state,
+		String postalCode,
 		String city,
-		String state
+		String country
 		
 	) {
 		BooleanBuilder where = dynamicWhere(
 			
 			
-			
-			
-			
-			
 			employeesId,
 			
+			
+			
+			
+			
 			reviewsList,
-			postalCode,
-			addressLine1,
 			lastName,
-			creditLimit,
-			firstName,
-			country,
 			customerName,
 			phone,
 			addressLine2,
+			creditLimit,
+			addressLine1,
+			firstName,
+			state,
+			postalCode,
 			city,
-			state
+			country
 				
 		);
 		return customersRepository.findAll(where);
@@ -158,47 +158,47 @@ public class CustomersServiceImpl implements CustomersService {
 		Pageable page,
 		
 		
-		
-		
-		
-		
 		Long employeesId,
 		
+		
+		
+		
+		
 		List<Reviews> reviewsList,
-		String postalCode,
-		String addressLine1,
 		String lastName,
-		Double creditLimit,
-		String firstName,
-		String country,
 		String customerName,
 		String phone,
 		String addressLine2,
+		Double creditLimit,
+		String addressLine1,
+		String firstName,
+		String state,
+		String postalCode,
 		String city,
-		String state
+		String country
 		
 	) {
 		BooleanBuilder where = dynamicWhere(
 			
 			
-			
-			
-			
-			
 			employeesId,
 			
+			
+			
+			
+			
 			reviewsList,
-			postalCode,
-			addressLine1,
 			lastName,
-			creditLimit,
-			firstName,
-			country,
 			customerName,
 			phone,
 			addressLine2,
+			creditLimit,
+			addressLine1,
+			firstName,
+			state,
+			postalCode,
 			city,
-			state
+			country
 			
 		);
 		return customersRepository.findAll(where, page);
@@ -207,24 +207,24 @@ public class CustomersServiceImpl implements CustomersService {
 	public BooleanBuilder dynamicWhere(
 		
 		
-		
-		
-		
-		
 		Long employeesId,
 		
+		
+		
+		
+		
 		List<Reviews> reviewsList,
-		String postalCode,
-		String addressLine1,
 		String lastName,
-		Double creditLimit,
-		String firstName,
-		String country,
 		String customerName,
 		String phone,
 		String addressLine2,
+		Double creditLimit,
+		String addressLine1,
+		String firstName,
+		String state,
+		String postalCode,
 		String city,
-		String state
+		String country
 		
 	) {
 		QCustomers qCustomers = QCustomers.customers;
@@ -233,34 +233,19 @@ public class CustomersServiceImpl implements CustomersService {
 	
 		
 		
-		
-		
-		
-		
 		if (employeesId != null) {
 			where.and(qCustomers.employees.id.eq(employeesId));
 		}
 		
+		
+		
+		
+		
 		if (reviewsList != null) {
 			where.and(qCustomers.reviewsList.any().in(reviewsList));
 		}
-		if (postalCode != null) {
-			where.and(qCustomers.postalCode.containsIgnoreCase(postalCode));
-		}
-		if (addressLine1 != null) {
-			where.and(qCustomers.addressLine1.containsIgnoreCase(addressLine1));
-		}
 		if (lastName != null) {
 			where.and(qCustomers.lastName.containsIgnoreCase(lastName));
-		}
-		if (creditLimit != null) {
-			where.and(qCustomers.creditLimit.eq(creditLimit));
-		}
-		if (firstName != null) {
-			where.and(qCustomers.firstName.containsIgnoreCase(firstName));
-		}
-		if (country != null) {
-			where.and(qCustomers.country.containsIgnoreCase(country));
 		}
 		if (customerName != null) {
 			where.and(qCustomers.customerName.containsIgnoreCase(customerName));
@@ -271,30 +256,52 @@ public class CustomersServiceImpl implements CustomersService {
 		if (addressLine2 != null) {
 			where.and(qCustomers.addressLine2.containsIgnoreCase(addressLine2));
 		}
-		if (city != null) {
-			where.and(qCustomers.city.containsIgnoreCase(city));
+		if (creditLimit != null) {
+			where.and(qCustomers.creditLimit.eq(creditLimit));
+		}
+		if (addressLine1 != null) {
+			where.and(qCustomers.addressLine1.containsIgnoreCase(addressLine1));
+		}
+		if (firstName != null) {
+			where.and(qCustomers.firstName.containsIgnoreCase(firstName));
 		}
 		if (state != null) {
 			where.and(qCustomers.state.containsIgnoreCase(state));
+		}
+		if (postalCode != null) {
+			where.and(qCustomers.postalCode.containsIgnoreCase(postalCode));
+		}
+		if (city != null) {
+			where.and(qCustomers.city.containsIgnoreCase(city));
+		}
+		if (country != null) {
+			where.and(qCustomers.country.containsIgnoreCase(country));
 		}
 		
 	
 		return where;
 	}
 	
-	
-	
-	
-	
-	
-	
-	
-	
-	
 	@Override
 	public void deleteOneById(Long id) {
-		customersRepository.deleteById(id);
+		Customers currentCustomers = customersRepository.getOne(id);
+		Iterable<Customers> listOfCustomers = customersRepository.findByEmployeesIdEquals(currentCustomers.getEmployees().getId());
+		Long size = listOfCustomers.spliterator().getExactSizeIfKnown();
+		if (size == 1) {
+		    return;
+		} else {
+		    customersRepository.deleteById(id);
+		}
 	}
+	
+	
+	
+	
+	
+	
+	
+	
+
