+++ ClassicCarsSpringBoot/src/main/java/com/vladolium/odinmodel/serviceImpl/OrderDetailsServiceImpl.java
@@ -59,6 +59,11 @@ public class OrderDetailsServiceImpl implements OrderDetailsService {
 	}
 	
 	
+	
+	
+	
+	
+	
 	@Override
 	public Iterable<OrderDetails> readAllByProductsId(Long productsId) {
 		return orderDetailsRepository.findByProductsIdEquals(productsId);
@@ -68,11 +73,6 @@ public class OrderDetailsServiceImpl implements OrderDetailsService {
 	public Page<OrderDetails> readAllByProductsId(Long productsId, Pageable page) {
 		return orderDetailsRepository.findByProductsIdEquals(productsId, page);
 	}
-	
-	
-	
-	
-	
 	
 	
 	
@@ -97,32 +97,32 @@ public class OrderDetailsServiceImpl implements OrderDetailsService {
 	@Override
 	public Iterable<OrderDetails> search(
 		
+		
+		
+		
+		
+		
 		Long productsId,
 		
 		
-		
-		
-		
-		
-		
-		Integer quantityOrdered,
+		Integer orderLineNumber,
 		Double priceEach,
-		Integer orderLineNumber
+		Integer quantityOrdered
 		
 	) {
 		BooleanBuilder where = dynamicWhere(
 			
+			
+			
+			
+			
+			
 			productsId,
 			
 			
-			
-			
-			
-			
-			
-			quantityOrdered,
+			orderLineNumber,
 			priceEach,
-			orderLineNumber
+			quantityOrdered
 				
 		);
 		return orderDetailsRepository.findAll(where);
@@ -132,32 +132,32 @@ public class OrderDetailsServiceImpl implements OrderDetailsService {
 	public Page<OrderDetails> searchPagination(
 		Pageable page,
 		
+		
+		
+		
+		
+		
 		Long productsId,
 		
 		
-		
-		
-		
-		
-		
-		Integer quantityOrdered,
+		Integer orderLineNumber,
 		Double priceEach,
-		Integer orderLineNumber
+		Integer quantityOrdered
 		
 	) {
 		BooleanBuilder where = dynamicWhere(
 			
+			
+			
+			
+			
+			
 			productsId,
 			
 			
-			
-			
-			
-			
-			
-			quantityOrdered,
+			orderLineNumber,
 			priceEach,
-			orderLineNumber
+			quantityOrdered
 			
 		);
 		return orderDetailsRepository.findAll(where, page);
@@ -165,17 +165,17 @@ public class OrderDetailsServiceImpl implements OrderDetailsService {
 	
 	public BooleanBuilder dynamicWhere(
 		
+		
+		
+		
+		
+		
 		Long productsId,
 		
 		
-		
-		
-		
-		
-		
-		Integer quantityOrdered,
+		Integer orderLineNumber,
 		Double priceEach,
-		Integer orderLineNumber
+		Integer quantityOrdered
 		
 	) {
 		QOrderDetails qOrderDetails = QOrderDetails.orderDetails;
@@ -183,24 +183,24 @@ public class OrderDetailsServiceImpl implements OrderDetailsService {
 		BooleanBuilder where = new BooleanBuilder();
 	
 		
+		
+		
+		
+		
+		
 		if (productsId != null) {
 			where.and(qOrderDetails.products.id.eq(productsId));
 		}
 		
 		
-		
-		
-		
-		
-		
-		if (quantityOrdered != null) {
-			where.and(qOrderDetails.quantityOrdered.eq(quantityOrdered));
+		if (orderLineNumber != null) {
+			where.and(qOrderDetails.orderLineNumber.eq(orderLineNumber));
 		}
 		if (priceEach != null) {
 			where.and(qOrderDetails.priceEach.eq(priceEach));
 		}
-		if (orderLineNumber != null) {
-			where.and(qOrderDetails.orderLineNumber.eq(orderLineNumber));
+		if (quantityOrdered != null) {
+			where.and(qOrderDetails.quantityOrdered.eq(quantityOrdered));
 		}
 		
 	
@@ -213,7 +213,14 @@ public class OrderDetailsServiceImpl implements OrderDetailsService {
 	
 	@Override
 	public void deleteOneById(Long id) {
-		orderDetailsRepository.deleteById(id);
+		OrderDetails currentOrderDetails = orderDetailsRepository.getOne(id);
+		Iterable<OrderDetails> listOfOrderDetails = orderDetailsRepository.findByProductsIdEquals(currentOrderDetails.getProducts().getId());
+		Long size = listOfOrderDetails.spliterator().getExactSizeIfKnown();
+		if (size == 1) {
+		    return;
+		} else {
+		    orderDetailsRepository.deleteById(id);
+		}
 	}
